Added informative messages to the build script build.sh.

+++ build.sh
@@ -14,7 +14,10 @@ PPC_HOST=$1
 
 HERE=$(pwd)
 
-rm -fr /tmp/ready-lisp
+if [ -d /tmp/ready-lisp ]; then
+    echo Removing previous ready-lisp build ...
+    rm -fr /tmp/ready-lisp
+fi
 
 if [[ -n "$PPC_HOST" ]]; then
     ssh $PPC_HOST rm -fr /tmp/ready-lisp || exit 1
@@ -23,8 +26,10 @@ fi
 cd /tmp
 
 if [[ $local == true ]]; then
+    echo Copying source tree to /tmp/ready-lisp ...
     rsync -a ~/src/ready-lisp . || exit 1
 else
+    echo Cloning ready-lisp from GitHub to /tmp/ready-lisp ...
     git clone git://github.com/jwiegley/ready-lisp.git || exit 1
 fi
 
@@ -34,6 +39,9 @@ else
     cd ready-lisp && make || exit 1
 fi
 
+echo Moving final disk image back to source tree ...
 mv build/ReadyLisp-*.dmg $HERE
+
+echo Ready Lisp build complete.
 
 # build.sh ends here
